--- a/packages/payload/src/checkPayloadDependencies.ts
+++ b/packages/payload/src/checkPayloadDependencies.ts
@@ -1,7 +1,7 @@
 import { checkDependencies } from './utilities/dependencies/dependencyChecker.js'
 import { PAYLOAD_PACKAGE_LIST } from './versions/payloadPackageList.js'
 
-export async function checkPayloadDependencies() {
+export function checkPayloadDependencies() {
   const dependencies = [...PAYLOAD_PACKAGE_LIST]
 
   if (process.env.PAYLOAD_CI_DEPENDENCY_CHECKER !== 'true') {
@@ -9,7 +9,7 @@ export async function checkPayloadDependencies() {
   }
 
   // First load. First check if there are mismatching dependency versions of payload packages
-  await checkDependencies({
+  void checkDependencies({
     dependencyGroups: [
       {
         name: 'payload',