--- a/packages/payload/src/collections/config/client.ts
+++ b/packages/payload/src/collections/config/client.ts
@@ -9,10 +9,10 @@ import type {
 } from '../../config/types.js'
 import type { ClientField } from '../../fields/config/client.js'
 import type { Payload } from '../../types/index.js'
+import type { SanitizedUploadConfig } from '../../uploads/types.js'
 import type { SanitizedCollectionConfig } from './types.js'
 
 import { createClientFields } from '../../fields/config/client.js'
-import { deepCopyObjectSimple } from '../../utilities/deepCopyObject.js'
 
 export type ServerOnlyCollectionProperties = keyof Pick<
   SanitizedCollectionConfig,
@@ -47,12 +47,16 @@ export type ClientCollectionConfig = {
     | 'preview'
     | ServerOnlyCollectionAdminProperties
   >
+  auth?: Omit<SanitizedCollectionConfig['auth'], 'forgotPassword' | 'strategies' | 'verify'>
   fields: ClientField[]
   labels: {
     plural: StaticLabel
     singular: StaticLabel
   }
-} & Omit<SanitizedCollectionConfig, 'admin' | 'fields' | 'labels' | ServerOnlyCollectionProperties>
+} & Omit<
+  SanitizedCollectionConfig,
+  'admin' | 'auth' | 'fields' | 'labels' | ServerOnlyCollectionProperties
+>
 
 const serverOnlyCollectionProperties: Partial<ServerOnlyCollectionProperties>[] = [
   'hooks',
@@ -92,97 +96,144 @@ export const createClientCollectionConfig = ({
   i18n: I18nClient
   importMap: ImportMap
 }): ClientCollectionConfig => {
-  const clientCollection = deepCopyObjectSimple(
-    collection,
-    true,
-  ) as unknown as ClientCollectionConfig
+  const clientCollection = {} as Partial<ClientCollectionConfig>
 
-  clientCollection.fields = createClientFields({
-    clientFields: clientCollection?.fields || [],
-    defaultIDType,
-    fields: collection.fields,
-    i18n,
-    importMap,
-  })
-
-  serverOnlyCollectionProperties.forEach((key) => {
-    if (key in clientCollection) {
-      delete clientCollection[key]
+  for (const key in collection) {
+    if (serverOnlyCollectionProperties.includes(key as any)) {
+      continue
     }
-  })
-
-  if ('upload' in clientCollection && typeof clientCollection.upload === 'object') {
-    serverOnlyUploadProperties.forEach((key) => {
-      if (key in clientCollection.upload) {
-        delete clientCollection.upload[key]
-      }
-    })
-
-    if ('imageSizes' in clientCollection.upload && clientCollection.upload.imageSizes.length) {
-      clientCollection.upload.imageSizes = clientCollection.upload.imageSizes.map((size) => {
-        const sanitizedSize = { ...size }
-        if ('generateImageName' in sanitizedSize) {
-          delete sanitizedSize.generateImageName
+    switch (key) {
+      case 'admin':
+        if (!collection.admin) {
+          break
         }
-        return sanitizedSize
-      })
+        clientCollection.admin = {} as ClientCollectionConfig['admin']
+        for (const adminKey in collection.admin) {
+          if (serverOnlyCollectionAdminProperties.includes(adminKey as any)) {
+            continue
+          }
+
+          switch (adminKey) {
+            case 'description':
+              if (
+                typeof collection.admin.description === 'string' ||
+                typeof collection.admin.description === 'object'
+              ) {
+                if (collection.admin.description) {
+                  clientCollection.admin.description = collection.admin.description
+                }
+              } else if (typeof collection.admin.description === 'function') {
+                const description = collection.admin.description({ t: i18n.t })
+                if (description) {
+                  clientCollection.admin.description = description
+                }
+              }
+              break
+            case 'livePreview':
+              clientCollection.admin.livePreview =
+                {} as ClientCollectionConfig['admin']['livePreview']
+              if (collection.admin.livePreview.breakpoints) {
+                clientCollection.admin.livePreview.breakpoints =
+                  collection.admin.livePreview.breakpoints
+              }
+              break
+            case 'preview':
+              if (collection.admin.preview) {
+                clientCollection.admin.preview = true
+              }
+              break
+            default:
+              clientCollection.admin[adminKey] = collection.admin[adminKey]
+          }
+        }
+        break
+      case 'auth':
+        if (!clientCollection.auth) {
+          break
+        }
+        clientCollection.auth = {} as SanitizedCollectionConfig['auth']
+        if (collection.auth.cookies) {
+          clientCollection.auth.cookies = collection.auth.cookies
+        }
+        if (collection.auth.depth !== undefined) {
+          // Check for undefined as it can be a number (0)
+          clientCollection.auth.depth = collection.auth.depth
+        }
+        if (collection.auth.disableLocalStrategy) {
+          clientCollection.auth.disableLocalStrategy = collection.auth.disableLocalStrategy
+        }
+        if (collection.auth.lockTime !== undefined) {
+          // Check for undefined as it can be a number (0)
+          clientCollection.auth.lockTime = collection.auth.lockTime
+        }
+        if (collection.auth.loginWithUsername) {
+          clientCollection.auth.loginWithUsername = collection.auth.loginWithUsername
+        }
+        if (collection.auth.maxLoginAttempts !== undefined) {
+          // Check for undefined as it can be a number (0)
+          clientCollection.auth.maxLoginAttempts = collection.auth.maxLoginAttempts
+        }
+        if (collection.auth.removeTokenFromResponses) {
+          clientCollection.auth.removeTokenFromResponses = collection.auth.removeTokenFromResponses
+        }
+
+        if (collection.auth.useAPIKey) {
+          clientCollection.auth.useAPIKey = collection.auth.useAPIKey
+        }
+        if (collection.auth.useAPIKey) {
+          clientCollection.auth.tokenExpiration = collection.auth.tokenExpiration
+        }
+        break
+      case 'fields':
+        clientCollection.fields = createClientFields({
+          defaultIDType,
+          fields: collection.fields,
+          i18n,
+          importMap,
+        })
+        break
+      case 'labels':
+        clientCollection.labels = {
+          plural:
+            typeof collection.labels.plural === 'function'
+              ? collection.labels.plural({ t: i18n.t })
+              : collection.labels.plural,
+          singular:
+            typeof collection.labels.singular === 'function'
+              ? collection.labels.singular({ t: i18n.t })
+              : collection.labels.singular,
+        }
+        break
+      case 'upload':
+        if (!collection.upload) {
+          break
+        }
+        clientCollection.upload = {} as SanitizedUploadConfig
+        for (const uploadKey in collection.upload) {
+          if (serverOnlyUploadProperties.includes(uploadKey as any)) {
+            continue
+          }
+          if (uploadKey === 'imageSizes') {
+            clientCollection.upload.imageSizes = collection.upload.imageSizes.map((size) => {
+              const sanitizedSize = { ...size }
+              if ('generateImageName' in sanitizedSize) {
+                delete sanitizedSize.generateImageName
+              }
+              return sanitizedSize
+            })
+          } else {
+            clientCollection.upload[uploadKey] = collection.upload[uploadKey]
+          }
+        }
+        break
+
+        break
+      default:
+        clientCollection[key] = collection[key]
     }
   }
 
-  if ('auth' in clientCollection && typeof clientCollection.auth === 'object') {
-    delete clientCollection.auth.strategies
-    delete clientCollection.auth.forgotPassword
-    delete clientCollection.auth.verify
-  }
-
-  if (collection.labels) {
-    Object.entries(collection.labels).forEach(([labelType, collectionLabel]) => {
-      if (typeof collectionLabel === 'function') {
-        clientCollection.labels[labelType] = collectionLabel({ t: i18n.t })
-      }
-    })
-  }
-
-  if (!clientCollection.admin) {
-    clientCollection.admin = {} as ClientCollectionConfig['admin']
-  }
-
-  serverOnlyCollectionAdminProperties.forEach((key) => {
-    if (key in clientCollection.admin) {
-      delete clientCollection.admin[key]
-    }
-  })
-
-  if (collection.admin.preview) {
-    clientCollection.admin.preview = true
-  }
-
-  let description = undefined
-
-  if (collection.admin?.description) {
-    if (
-      typeof collection.admin?.description === 'string' ||
-      typeof collection.admin?.description === 'object'
-    ) {
-      description = collection.admin.description
-    } else if (typeof collection.admin?.description === 'function') {
-      description = collection.admin?.description({ t: i18n.t })
-    }
-  }
-
-  if (description) {
-    clientCollection.admin.description = description
-  }
-
-  if (
-    'livePreview' in clientCollection.admin &&
-    clientCollection.admin.livePreview &&
-    'url' in clientCollection.admin.livePreview
-  ) {
-    delete clientCollection.admin.livePreview.url
-  }
-
-  return clientCollection
+  return clientCollection as ClientCollectionConfig
 }
 
 export const createClientCollectionConfigs = ({