--- a/packages/ui/src/utilities/buildFieldSchemaMap/traverseFields.ts
+++ b/packages/ui/src/utilities/buildFieldSchemaMap/traverseFields.ts
@@ -63,7 +63,6 @@ export const traverseFields = ({
 
         break
       case 'collapsible':
-
       case 'row':
         traverseFields({
           config,